feat: replace Lemon Squeezy with usage-based pricing model
Replace the fixed 25 AED/assessment Lemon billing with usage-based
pricing in micro-credits ($0.000001 per credit). Customers pay for actual
Claude tokens consumed plus a feature-specific markup; new orgs get $1.50
of free credits on signup.

Key changes:
- New pricing_service: 3x markup on score/assessment, 1x on prescreen,
  0.10x on cache hit. Free tier $1.50, top-up packs $20/$100/$500.
- usage_metering_service writes usage_events + ledger debits with shadow
  vs live mode (USAGE_METER_LIVE flag, default off).
- claude_client_resolver provisions per-org Anthropic Workspace keys
  lazily via Admin API; falls back to shared Taali key on any failure.
- Stripe one-time-payment top-up flow replaces Lemon checkout.
- Settings billing tab fully rebuilt: balance, packs, per-feature
  breakdown, usage events log. Landing page gets a pricing section.
- Lemon Squeezy package, settings, and migrations fully removed.
- Migration 052 adds usage_events + usage_grants + workspace key columns
  + bigint conversion. Migration 053 drops legacy assessments_used /
  assessments_limit.

Memory: anthropic_key_routing.md updated to reflect per-org workspace
keys (still Taali-owned, Admin-API-provisioned).

Verified: 1141 backend tests pass, frontend vite build succeeds.

+++ backend/app/models/billing_credit_ledger.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, JSON, String
+from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, Integer, JSON, String
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
@@ -10,8 +10,10 @@ class BillingCreditLedger(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     organization_id = Column(Integer, ForeignKey("organizations.id"), index=True, nullable=False)
-    delta = Column(Integer, nullable=False)
-    balance_after = Column(Integer, nullable=False)
+    # delta and balance_after are micro-credits as of the 2026-04-29 usage-
+    # based pricing migration. Pre-migration rows used whole-credit Integer.
+    delta = Column(BigInteger, nullable=False)
+    balance_after = Column(BigInteger, nullable=False)
     reason = Column(String, nullable=False)
     external_ref = Column(String, nullable=True, unique=True, index=True)
     assessment_id = Column(Integer, ForeignKey("assessments.id"), index=True, nullable=True)
